IDE0079: Remove unnecessary suppression SA1649
